move personal scripts out of main, jsonignore getgit in repo, generate folder structure before get
