+++ src/main/java/org/example/fetching/CachedDataRepoFetcher.java
@@ -44,15 +44,15 @@ public class CachedDataRepoFetcher {
             return mapper.readValue(output, Repository.class);
         }
 
-        logger.debug("{} not found or forced update, getting from GitHub.", repoName);
-        Repository repo = new Repository(gh.getRepository(repoName));
-
-        logger.info("Done reading from API, writing to file.");
-
         if(!output.getParentFile().exists() && !output.getParentFile().mkdirs()) {
             logger.error("Failed to create directories, necessary to save repo data.");
             throw new RemoteException("Error `.mkdirs()`. Directories not created.");
         }
+
+        logger.debug("{} not found or forced update, getting from GitHub.", repoName);
+        Repository repo = new Repository(gh.getRepository(repoName));
+
+        logger.info("Done reading from API, writing to file.");
 
         mapper.writerWithDefaultPrettyPrinter()
                 .writeValue(output, repo);
